Chore #11: feature package align
