+++ app/src/main/java/org/sopt/and/feature/home/HomeState.kt
@@ -1,8 +1,8 @@
-package org.sopt.and.home
+package org.sopt.and.feature.home
 
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.persistentListOf
-import org.sopt.and.home.model.ContentModel
+import org.sopt.and.feature.home.model.ContentModel
 
 data class HomeState(
     val topBannerContent: PersistentList<ContentModel> = persistentListOf(),
