+++ app/src/main/java/org/sopt/and/feature/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.main
+package org.sopt.and.feature.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
