+++ app/src/main/java/org/sopt/and/feature/my/MyNavigation.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.search
+package org.sopt.and.feature.my
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,21 +6,23 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import kotlinx.serialization.Serializable
-import org.sopt.and.main.MainTabRoute
+import org.sopt.and.feature.main.MainTabRoute
 
-fun NavController.navigateSearch(
+fun NavController.navigateMy(
     navOptions: NavOptions
 ) {
-    navigate(Search, navOptions)
+    navigate(My, navOptions)
 }
 
-fun NavGraphBuilder.searchNavGraph(
+fun NavGraphBuilder.myNavGraph(
     paddingValues: PaddingValues,
 ) {
-    composable<Search> {
-        SearchRoute(paddingValues = paddingValues)
+    composable<My> {
+        MyRoute(
+            paddingValues = paddingValues,
+        )
     }
 }
 
 @Serializable
-data object Search : MainTabRoute
+data object My : MainTabRoute
