+++ app/src/main/java/org/sopt/and/feature/search/SearchRoute.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.search
+package org.sopt.and.feature.search
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.runtime.Composable
