+++ app/src/main/java/org/sopt/and/feature/signup/SignUpNavigation.kt
@@ -1,10 +1,10 @@
-package org.sopt.and.signup
+package org.sopt.and.feature.signup
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import kotlinx.serialization.Serializable
-import org.sopt.and.main.MainTabRoute
+import org.sopt.and.feature.main.MainTabRoute
 
 fun NavController.navigateSignUp() {
     navigate(SignUp)
