--- app/src/main/java/org/sopt/and/signin/SignInState.kt
@@ -1,9 +1,0 @@
-package org.sopt.and.signin
-
-data class SignInState(
-    val email: String = "",
-    val password: String = "",
-    var isPasswordVisible: Boolean = false,
-) {
-    val isButtonEnabled: Boolean = email.isNotEmpty() && password.isNotEmpty()
-}
