--- app/src/main/java/org/sopt/and/signup/SignUpState.kt
@@ -1,9 +1,0 @@
-package org.sopt.and.signup
-
-data class SignUpState(
-    val email: String = "",
-    val password: String = "",
-    var isPasswordVisible: Boolean = false,
-){
-    val isButtonEnabled : Boolean = email.isNotEmpty() && password.isNotEmpty()
-}
